+++ workers/tests/unittests/test_api_request_models.py
@@ -1,7 +1,9 @@
 """Test module for the `pydantic` data models for API requests"""
-
+import json
 import os
 import unittest
+
+from geojson import Polygon
 
 from ohsome_quality_analyst.api import request_models
 
@@ -14,19 +16,15 @@ class TestApiRequestModels(unittest.TestCase):
             "heidelberg-altstadt-feature.geojson",
         )
         with open(path, "r") as file:
-            self.bpolys = file.read()
+            self.bpolys = json.load(file)
 
     def test_bpolys_valid(self):
         request_models.BaseRequestModel(bpolys=self.bpolys)
 
     def test_bpolys_invalid(self):
-        path = os.path.join(
-            os.path.dirname(os.path.abspath(__file__)),
-            "fixtures",
-            "ohsome-response-200-invalid.geojson",
+        bpolys = Polygon(
+            [[(2.38, 57.322), (23.194, -20.28), (-120.43, 19.15), (2.0, 1.0)]]
         )
-        with open(path, "r") as file:
-            bpolys = file.read()
 
         with self.assertRaises(ValueError):
             request_models.BaseRequestModel(bpolys=bpolys)