+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking Changes
 
 - Change API Path parameter `name` to be a Query parameter instead of a Path parameter ([#190])
+- Change type of API parameter `bpolys` for POST requests to JSON (`dict`) instead of string ([#204])
 
 ### Other Changes
 
@@ -15,9 +16,11 @@
 
 - Update requests for `/indicator` and `/report` endpoints. `name` is now a *Query* parameter ([#190])
     - E.g. `/indicator/GhsPopComparisonBuilding` -> `indicator?name=GhsPopComparisonBuilding`
+- Update `bpolys` parameter of POST requests to be a GeoJSON object instead of a string ([#204])
 
 [#190]: https://github.com/GIScience/ohsome-quality-analyst/pull/190
 [#199]: https://github.com/GIScience/ohsome-quality-analyst/pull/199
+[#204]: https://github.com/GIScience/ohsome-quality-analyst/pull/204
 
 
 ## 0.6.0